pyproject.toml
[project]
name = "overwatch"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "celery[redis]>=5.5.2",
    "clip",
    "django>=5.2.1",
    "django-celery-results>=2.6.0",
    "numpy>=2.2.5",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.9",
]

[tool.ruff.lint]
extend-select = ["I", "F401"]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }
